feat[user-service]: implement interest related endpoints and fix some issue with previous endpoints

+++ backend/user-service/src/main/java/org/bitmonsters/userservice/interest/Interest.java
@@ -17,12 +17,12 @@ public class Interest {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(nullable = false)
     private User user;
 
     @Column(nullable = false)
-    private Long tagId;
+    private Integer tagId;
 }
